fix(revive): fix revive linter warnings
